+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Glide is now responsive when turning (finally T.T).
 * Heartless and Nobody spawns in the end:
   * They are less frequent in the main island while they keep being frequent in the outer islands.
+* Armor items will now display their stats in the Shop screen.
 
 **Fixed:**
 * Music discs not having the duration nor the track name in the tooltip.